refactor:  useMermaid
