+++ projects/portfolio/src/client/components/pages/Chat.tsx
@@ -1,8 +1,6 @@
 import { hc } from 'hono/client'
-import mermaid from 'mermaid'
 import React, {
   useEffect,
-  useCallback,
   useRef,
   useState,
   useMemo,
@@ -21,6 +19,7 @@ import type { AppType } from '@/server/app.d'
 import { ChatInput } from '@/client/components/input/ChatInput'
 import { FileImageInput, FileImagePreview } from '@/client/components/input/FileImageInput'
 import { ToggleInput } from '@/client/components/input/ToggleInput'
+import { useMermaid } from '@/client/components/markdown/useMermaid'
 import { ArrowUpIcon } from '@/client/components/svg/ArrowUpIcon'
 import { ChatbotIcon } from '@/client/components/svg/ChatbotIcon'
 import { CheckIcon } from '@/client/components/svg/CheckIcon'
@@ -158,19 +157,19 @@ type MarkdownCodeBlockProps = React.HTMLAttributes<HTMLElement> & {
   children?: React.ReactNode | string
 }
 
-const randomid = () => Number.parseInt(String(Math.random() * 1e15), 10).toString(36)
-
 function MarkdownCodeBlock({ className, children }: MarkdownCodeBlockProps) {
   const [copied, setCopied] = useState(false)
-  const demoid = useRef(`dome${randomid()}`)
-  const [container, setContainer] = useState<HTMLDivElement | null>(null)
   const code =
     typeof children === 'string' ? children : Array.isArray(children) ? children.join('') : ''
+
   const language = className?.split('-')[1]
-  const isMermaid = className && /^language-mermaid/.test(className.toLocaleLowerCase())
+
+  const { isMermaid, mermaidRef } = useMermaid(code, className)
+
   if (typeof children !== 'string' || !language) {
     return <code>{children}</code>
   }
+
   const handleClickCopy = async () => {
     setCopied(true)
     try {
@@ -181,26 +180,6 @@ function MarkdownCodeBlock({ className, children }: MarkdownCodeBlockProps) {
     }
     setCopied(false)
   }
-  const reRender = async () => {
-    if (container && isMermaid) {
-      try {
-        const str = await mermaid.render(demoid.current, code)
-        container.innerHTML = str.svg
-      } catch (error) {
-        container.innerHTML = String(error)
-      }
-    }
-  }
-
-  useEffect(() => {
-    reRender()
-  }, [container, isMermaid, code, demoid])
-
-  const refElement = useCallback((node: HTMLDivElement | null) => {
-    if (node !== null) {
-      setContainer(node)
-    }
-  }, [])
 
   return (
     <>
@@ -226,16 +205,8 @@ function MarkdownCodeBlock({ className, children }: MarkdownCodeBlockProps) {
       </div>
       <SyntaxHighlighter language={language}>{children}</SyntaxHighlighter>
       {isMermaid && (
-        <div className='flex justify-end'>
-          <button type='button' className='flex cursor-pointer gap-1 align-center'>
-            <CheckIcon size={18} className='stroke-white' />
-            <span className='text-xs'>プレビュー表示</span>
-          </button>
-        </div>
-      )}
-      {isMermaid && (
-        <div className='my-1 bg-white p-2'>
-          <code ref={refElement} data-name='mermaid' />
+        <div className='mt-2 rounded-md bg-white p-2'>
+          <code ref={mermaidRef} data-name='mermaid' />
         </div>
       )}
     </>
